pyproject.toml
[tool.poetry]
name = "myapp"
version = "0.1.0"
description = ""
authors = ["Clement Ke"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.21.1"
fastapi = "^0.95.1"
fastapi-versioning = "^0.10.0"
pydantic = {extras = ["email"], version = "^1.10.7"}
pydantic-mongo = "^1.0.1"
pymongo = "^4.3.3"
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-keycloak = "^2.16.0"
authlib = "^1.2.0"
aiohttp = "^3.8.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
httpx = "^0.24.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]